del sample-mapper. rename sample2-mapper to sample-mapper
